+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @guanxiangkai/ui
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @guanxiangkai/platform-client@0.3.0
+
 ## 0.2.0
 
 ### Minor Changes